pyproject.toml
[project]
name = "tokenizer-workshop"
version = "0.1.0"
description = "An educational Python project for understanding tokenization and building tokenizers step by step."
readme = "README.md"
authors = [{ name = "Burak" }]
requires-python = ">=3.10"
dependencies = [
    "groq>=1.1.2",
    "pyyaml>=6.0.3",
]

[project.scripts]
tokenizer-workshop = "tokenizer_workshop:main"

[build-system]
requires = ["uv_build>=0.11.6,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
